// src/generated/prisma.ts
interface Node {
  """The id of the object."""
  id: ID!
}
// ...
export type Role =   'USER' |
  'ADMIN'
// ...
export interface User extends Node {
  id: ID_Output
  email?: String
  emailVerified?: Boolean
  auth0Id?: String
  role?: Role
  name?: String
  avatar?: String
}

export interface AggregateUser {
  count: Int
}

/*
 * An edge in a connection.

 */
export interface UserEdge {
  node: User
  cursor: String
}

/*
The `String` scalar type represents textual data, represented as UTF-8 character sequences. The String type is most often used by GraphQL to represent free-form human-readable text.
*/
export type String = string

/*
The `Boolean` scalar type represents `true` or `false`.
*/
export type Boolean = boolean
// ...
export type ID_Output = string